[remix] Fix support for optional path params
Changes the way that routes with optional params are stored on Vercel to
not use `?` in the name, because it's not possible to have a route
destination with a `?` that's pointing to a function.

So instead of the function with a name such as `:lang?`, it will be
stored on Vercel with a name like `(:lang)`, which can be routed to.