[all] Use correct "license" field in `package.json`
Follow-up to #9696. Ensures that all packages in this repository
have the correct "license" field in their respective `package.json`
files set to "Apache-2.0", to match the `LICENSE` file at the root
of the repository.

The `LICENSE` file the authoritative source, so the "MIT" value
was incorrect. This change corrects the previous values and is not
considered a relicensing.