[cli] Remove `--target` global common arg
It doesn't make sense to have `--target` as a global argument, because
not all subcommands utilize this flag (actually only `vc deploy` does,
and it's even deprecated).

This is technically a breaking change (though risk of someone using this
flag on non-`vc deploy` is very unlikely), so we may want to hold off on
merging this until we're ready to do a major version bump.