[node] fix decompress mismatching
This PR disabled node-fetch default compression handling.

The default behavior in node-fetch is to handle the compression. That's great if you interact with node-fetch as user, but bad if you use it as intermediate proxy server, since it's creating a mismatching related with the body format and length expectations.

Instead, we disable compression, get the buffered response and compress it again if needed.