[frameworks][fs-detectors] Add "supercedes" prop to Framework interface
The `supercedes` property on the `Framework` interface may be specified
by a framework in which the superceded framework should _not_ be
included in the resulting framework matches during detection.

For example, a Hydrogen v2 project matches both the "remix" and
"hydrogen" framework detectors, but we really only want "remix" to be
selected.