Enable changesets experimental `updateInternalDependents: 'always'` config
From the sounds of it, setting this config value to "always" will make
it so that a patch release will happen for a package that depends on
another package which is being bumped. This would normally happen
anyways, but does not happen when a dependency is a `devDependency`
rather than a `dependency`.