[client] More helpful error message when `vc deploy --prebuilt` has missing files
Print a more helpful error message for the scenario the user encountered here: https://github.com/orgs/vercel/discussions/5612

What was happening in they were running `vc build` in one GH Action job,
and then `vc deploy --prebuilt` in a different job. The later job was
does not have `node_modules` populated, causing the deploy command to
fail.

The updated error message hints to the user that the `node_modules`
should be installed for the command to succeed.