[remix] Don't install Remix fork when not using split configuration
The fork of the Remix compiler is only necessary when the project is
utilizing split configurations in their routes (i.e. mixing Node.js and
Edge functions).

However, the injecting of the forked compiler has proven to be brittle
and has many edge cases which cause the fork to not be present.
Considering that the more common case is to _not_ use split
configuration, skip the fork entirely when that is the case.

This is somewhat of a band-aid / hold over for the "old" Remix. With
the Remix + Vite configuration, the fork will be entirely unnecessary.