[remix] Disable `prepareCache()` npm install for Remix + Vite
In legacy Remix apps, we were running `npm install` once again before
globbing the `node_modules` dir due to the injecting of the forked Remix
compiler which would cause a cache miss without that happening.

For Vite, we are no longer mucking about `node_modules` at build-time,
so there is no need to run `npm install` in the `prepareCache()`
function.

There is also no "cacheDirectory" setting so the globbing of the legacy
cache directory is also moved into the legacy-only logic.

tl;dr For Remix + Vite apps, the only thing that needs to be cached is
`node_modules`.