Move integration tests to Zero config team account (#11766)
After experiencing CI failures for our test account, it became evident that tests were being run against some older infrastructure. This PR moves the tests to a new account, and we should now be able to inspect build and runtime logs more easily for test failures

One of the biggest ramifications of this change is that many of our integration tests were operating under the assumption that we wanted to deploy to the default team for the token that was being used. However, there's no user whose default team is the newly created test account, so we've made some changes to explicitly pass the scope in all of our integration tests.