[redwood] Support API endpoint configuration via `export const config`
Uses the `@vercel/static-config` module to get the `export const config`
configuration for Redwood.js API endpoints, so it's the same interface
as vanilla JS endpoints, and Remix.

The previous usage of `getLambdaOptionsFromFunction` was broken because
it would cause the `@vercel/redwood` Builder to be executed multiple
times (once for each endpoint).