reject mismatched corepack and detected package managers (#11603)
When corepack is used and the package manager or version does not match the inferred package manager or version from the lockfile, we should throw an error. Right now, this can happen with `pnpm` (for example) where `pnpm` only logs a warning, causing indeterministic builds.

This PR mostly updates `getPathOverrideForPackageManager` to not care about corepack, allowing the calling code to compare the detected values to the corepack values.