[cli] Refactor common flags to single point of truth (#11936)
I noticed that several flags across the Vercel CLI are common in their long and short flag forms, including how those flags are parsed. This PR creates a single point of truth for a variety of common flags, some of which accidentally diverged in type prior to this change. 

Having a single point of truth makes it easier to keep these various places in sync.

There may be room for more of these common flags to be refactored similarly in the future.