[next] Support Request Chaining (#12117)
This enables the CLI to output the request chaining syntax used for
Partial Prerendering (PPR). This resolves a problem when the Next.js
application uses a `basePath` property as well as stabilizing the
feature in infrastructure.

Previously, the pathname (including the `basePath`) was used as both the
pathname and the output path. This resolves this by restoring ownership
of the header configuration to Next.js, which describes it in the
manifest.